ci: add istio to e2e
Refactors work submitted by @yashpal2104 to be more aligned with the structure of cert-manager and prometheus.
- Note I also rebased the branch off the latest upstream/notebooks-v2 branch.

Includes a commented out block of code that adds istio-injection label on the namespaces created by e2e.
- VirtualService objects must exist for this code to be uncommented - otherwise the Workspace connect endpoint is unreachable.

Moved handling of `istioctl` dependency to the `Makefile` as that code was unwiedly/intimidating to manage in a `.go` file.

Additionally, refactored utils module to be decomposed into multiple files as it was becoming unwieldy to manage everything in utils.go.
