+++ workspaces/controller/test/e2e/e2e_test.go
@@ -76,6 +76,14 @@ var _ = Describe("controller", Ordered, func() {
 		cmd = exec.Command("kubectl", "create", "ns", workspaceNamespace)
 		_, _ = utils.Run(cmd) // ignore errors because namespace may already exist
 
+		// TODO: enable Istio injection once we have logic to create VirtualServices during Workspace reconciliation
+		// By("labeling namespaces for Istio injection")
+		// err := utils.LabelNamespaceForIstioInjection(controllerNamespace)
+		// ExpectWithOffset(1, err).NotTo(HaveOccurred())
+
+		// err = utils.LabelNamespaceForIstioInjection(workspaceNamespace)
+		// ExpectWithOffset(1, err).NotTo(HaveOccurred())
+
 		By("creating common workspace resources")
 		cmd = exec.Command("kubectl", "apply",
 			"-k", filepath.Join(projectDir, "config/samples/common"),
