+++ workspaces/controller/test/utils/kind.go
@@ -1,0 +1,44 @@
+/*
+Copyright 2024.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package utils
+
+import (
+	"os"
+	"os/exec"
+	"strings"
+)
+
+// LoadImageToKindClusterWithName loads a local docker image to the kind cluster
+func LoadImageToKindClusterWithName(name string) error {
+	var cluster string
+	if v, ok := os.LookupEnv("KIND_CLUSTER"); ok {
+		cluster = v
+	} else {
+		// if `KIND_CLUSTER` is not set, get the cluster name from the kubeconfig
+		cmd := exec.Command("kubectl", "config", "current-context")
+		output, err := Run(cmd)
+		if err != nil {
+			return err
+		}
+		cluster = strings.TrimSpace(output)
+		cluster = strings.Replace(cluster, "kind-", "", 1)
+	}
+	kindOptions := []string{"load", "docker-image", name, "--name", cluster}
+	cmd := exec.Command("kind", kindOptions...)
+	_, err := Run(cmd)
+	return err
+}
